change the project folder structure, add webpack and babel and necessary dependencies
