+++ makeData.js
@@ -1,45 +1,44 @@
-import namor from 'namor'
+import namor from 'namor';
 
-const range = len => {
-  const arr = []
-  for (let i = 0; i < len; i++) {
-    arr.push(i)
+const range = (len) => {
+  const arr = [];
+  for (let i = 0; i < len; i += 1) {
+    arr.push(i);
   }
-  return arr
-}
+  return arr;
+};
 
 const newPerson = () => {
-  const statusChance = Math.random()
+  const statusChance = Math.random();
   return {
     firstName: namor.generate({ words: 1, numbers: 0 }),
     password: namor.generate({ words: 1, numbers: 0 }),
-    car:{
-      imgsrc:'imagesource',
-      name:'',
+    car: {
+      imgsrc: 'imagesource',
+      name: '',
     },
     lastName: namor.generate({ words: 1, numbers: 0 }),
     age: Math.floor(Math.random() * 30),
     visits: Math.floor(Math.random() * 100),
     progress: Math.floor(Math.random() * 100),
     status:
+      // eslint-disable-next-line no-nested-ternary
       statusChance > 0.66
         ? 'relationship'
         : statusChance > 0.33
-        ? 'complicated'
-        : 'single',
-  }
-}
+          ? 'complicated'
+          : 'single',
+  };
+};
 
 export default function makeData(...lens) {
   const makeDataLevel = (depth = 0) => {
-    const len = lens[depth]
-    return range(len).map(d => {
-      return {
-        ...newPerson(),
-        subRows: lens[depth + 1] ? makeDataLevel(depth + 1) : undefined,
-      }
-    })
-  }
+    const len = lens[depth];
+    return range(len).map(() => ({
+      ...newPerson(),
+      subRows: lens[depth + 1] ? makeDataLevel(depth + 1) : undefined,
+    }));
+  };
 
-  return makeDataLevel()
+  return makeDataLevel();
 }
