+++ package.json
@@ -40,11 +40,20 @@
     "webpack-dev-server": "^4.5.0"
   },
   "dependencies": {
+    "@emotion/react": "^11.6.0",
+    "@emotion/styled": "^11.6.0",
     "@mui/material": "^5.1.0",
+    "buffer": "^6.0.3",
+    "crypto": "^1.0.1",
+    "crypto-browserify": "^3.12.0",
+    "crypto-extra": "^1.0.1",
+    "crypto-js": "^4.1.1",
     "match-sorter": "^6.3.1",
+    "namor": "^2.0.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-table": "^7.7.0",
+    "stream-browserify": "^3.0.0",
     "styled-components": "^5.3.3"
   }
 }
