+++ src/components/DefaultColumnFilter.js
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const DefaultColumnFilter = ({
+  column: { filterValue, preFilteredRows, setFilter },
+}) => {
+  const count = preFilteredRows.length;
+
+  return (
+    <input
+      value={filterValue || ''}
+      onChange={(e) => {
+        setFilter(e.target.value || undefined); // Set undefined to remove the filter entirely
+      }}
+      placeholder={`Search Last Name ${count} records...`}
+    />
+  );
+};
+
+export default DefaultColumnFilter;
